app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.6.0
    # via pydantic
argon2-cffi==23.1.0
    # via -r requirements.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
blinker==1.7.0
    # via flask
cachetools==5.3.2
    # via google-auth
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
google-auth==2.23.4
    # via google-auth-oauthlib
google-auth-oauthlib==1.1.0
    # via -r requirements.in
greenlet==3.0.1
    # via sqlalchemy
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
oauthlib==3.2.2
    # via requests-oauthlib
psycopg2==2.9.9
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via pydantic-settings
requests==2.31.0
    # via requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
sqlalchemy==2.0.23
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.0.7
    # via requests
werkzeug==3.0.1
    # via flask
